Copy pasta ninja. Generating classes for @State elements
Generated classes do not depend on the enclosing class of the State
elements although it must for View subclasses since we should store the
parent state in the bundle, require it in the `saveState` method's
arguments, etc.

It looks like the BundleBinding should not be shared for State and Extra
elements since the reading logic should be a little different, most
notably, Extra does allow fallback on a method of the target for
providing a default value but State does not operate like that, whatever
was saved would be restored and no additional processing of the Bundle
should be performed, no checking for `containsKey` or whatnot. TBD.